refactor(store): remove execution strategy option
In this commit, we remove the execution strategy option, as discussed in one of
our internal meetings, because it isn't usable by the community and adds complexity
that isn't worth including in the public API.
