+++ packages/store/src/operators/leave-ngxs.ts
@@ -1,10 +1,11 @@
 import { ɵwrapObserverCalls } from '@ngxs/store/internals';
-import { NgxsExecutionStrategy } from '../execution/symbols';
+
+import { InternalNgxsExecutionStrategy } from '../execution/execution-strategy';
 
 /**
  * Returns operator that will run
  * `subscribe` outside of the ngxs execution context
  */
-export function leaveNgxs<T>(ngxsExecutionStrategy: NgxsExecutionStrategy) {
+export function leaveNgxs<T>(ngxsExecutionStrategy: InternalNgxsExecutionStrategy) {
   return ɵwrapObserverCalls<T>(fn => ngxsExecutionStrategy.leave(fn));
 }
