+++ packages/store/tests/dispatch.spec.ts
@@ -1,17 +1,11 @@
 import { ErrorHandler, Injectable, NgZone } from '@angular/core';
 import { fakeAsync, flush, TestBed, tick } from '@angular/core/testing';
-import {
-  State,
-  Action,
-  Store,
-  NgxsModule,
-  StateContext,
-  NgxsExecutionStrategy
-} from '@ngxs/store';
+import { State, Action, Store, NgxsModule, StateContext } from '@ngxs/store';
 import { of, throwError, timer } from 'rxjs';
 import { delay, map, tap } from 'rxjs/operators';
 
 import { NoopErrorHandler } from './helpers/utils';
+import { InternalNgxsExecutionStrategy } from '../src/execution/execution-strategy';
 
 describe('Dispatch', () => {
   class Increment {
@@ -38,17 +32,6 @@ describe('Dispatch', () => {
       }
     }
 
-    @Injectable({ providedIn: 'root' })
-    class FakeExecutionStrategy implements NgxsExecutionStrategy {
-      enter<T>(func: () => T): T {
-        return func();
-      }
-
-      leave<T>(func: () => T): T {
-        return func();
-      }
-    }
-
     @Injectable()
     class CustomErrorHandler implements ErrorHandler {
       handleError(error: Error): void {
@@ -58,11 +41,7 @@ describe('Dispatch', () => {
 
     // Act
     TestBed.configureTestingModule({
-      imports: [
-        NgxsModule.forRoot([CounterState], {
-          executionStrategy: FakeExecutionStrategy
-        })
-      ],
+      imports: [NgxsModule.forRoot([CounterState])],
       providers: [
         {
           provide: ErrorHandler,
@@ -78,13 +57,15 @@ describe('Dispatch', () => {
     let thrownMessage: typeof message | null = null;
 
     // Start spying after module is initialized, not before
-    jest.spyOn(FakeExecutionStrategy.prototype, 'leave').mockImplementationOnce(func => {
-      try {
-        return func();
-      } catch (e: any) {
-        thrownMessage = e.message;
-      }
-    });
+    jest
+      .spyOn(InternalNgxsExecutionStrategy.prototype, 'leave')
+      .mockImplementationOnce(func => {
+        try {
+          return func();
+        } catch (e: any) {
+          thrownMessage = e.message;
+        }
+      });
 
     store.dispatch(new Increment());
