+++ packages/store/tests/standalone-features/standalone-feature.spec.ts
@@ -1,19 +1,8 @@
 import { Component, Injectable, NgZone } from '@angular/core';
 import { bootstrapApplication } from '@angular/platform-browser';
 import { freshPlatform, skipConsoleLogging } from '@ngxs/store/internals/testing';
-import {
-  Action,
-  NoopNgxsExecutionStrategy,
-  State,
-  StateContext,
-  Store,
-  provideStates,
-  provideStore
-} from '@ngxs/store';
+import { Action, State, StateContext, Store, provideStates, provideStore } from '@ngxs/store';
 import { Router, provideRouter } from '@angular/router';
-
-import { NGXS_EXECUTION_STRATEGY } from '../../src/execution/symbols';
-import { DispatchOutsideZoneNgxsExecutionStrategy } from '../../src/execution/dispatch-outside-zone-ngxs-execution-strategy';
 
 describe('Standalone features', () => {
   class AddCountry {
@@ -32,64 +21,6 @@ describe('Standalone features', () => {
       ctx.setState(state => [...state, action.country]);
     }
   }
-
-  it(
-    'should bootstrap with no explicit options',
-    freshPlatform(async () => {
-      // Arrange
-      @Component({
-        selector: 'app-root',
-        template: '',
-        standalone: true
-      })
-      class TestComponent {}
-
-      // Act
-      const appRef = await skipConsoleLogging(() =>
-        bootstrapApplication(TestComponent, {
-          providers: [provideStore([])]
-        })
-      );
-
-      // Assert
-      expect(appRef.injector.get(NGXS_EXECUTION_STRATEGY)).toBeInstanceOf(
-        DispatchOutsideZoneNgxsExecutionStrategy
-      );
-
-      appRef.destroy();
-    })
-  );
-
-  it(
-    'should overwrite default values when `withNgxsConfig` is provided',
-    freshPlatform(async () => {
-      // Arrange
-      @Component({
-        selector: 'app-root',
-        template: '',
-        standalone: true
-      })
-      class TestComponent {}
-
-      // Act
-      const appRef = await skipConsoleLogging(() =>
-        bootstrapApplication(TestComponent, {
-          providers: [
-            provideStore([], {
-              executionStrategy: NoopNgxsExecutionStrategy
-            })
-          ]
-        })
-      );
-
-      // Assert
-      expect(appRef.injector.get(NGXS_EXECUTION_STRATEGY)).toBeInstanceOf(
-        NoopNgxsExecutionStrategy
-      );
-
-      appRef.destroy();
-    })
-  );
 
   it(
     'should bootstrap an app with `provideStore`',
